chore: fit tree-sitter 0.26 child() API
tree-sitter 0.26 changed Node::child(i) to take a u32 instead of
a usize. child_count was already cast to u32 a few lines above,
so the cast back to usize is now both redundant and a type error.

+++ src/analyzer.rs
@@ -1127,8 +1127,7 @@ impl<'a, 'f> Analyzer<'a, 'f> {
             let child_count = node.child_count() as u32;
             let first_child_kind_id = node.child(0).map_or(0, |c| c.kind_id());
             let last_child_kind_id = if child_count > 1 {
-                node.child(child_count as usize - 1)
-                    .map_or(0, |c| c.kind_id())
+                node.child(child_count - 1).map_or(0, |c| c.kind_id())
             } else {
                 first_child_kind_id
             };
